+++ TSS.Java/src/tss/tpm/TPM2_PolicyNV_REQUEST.java
@@ -21,13 +21,13 @@ public class TPM2_PolicyNV_REQUEST extends TpmStructure
      * @param _offset the octet offset in the NV Index for the start of operand A 
      * @param _operation the comparison to make
      */
-    public TPM2_PolicyNV_REQUEST(TPM_HANDLE _authHandle,TPM_HANDLE _nvIndex,TPM_HANDLE _policySession,byte[] _operandB,int _offset,TPM_EO _operation)
+    public TPM2_PolicyNV_REQUEST(TPM_HANDLE _authHandle,TPM_HANDLE _nvIndex,TPM_HANDLE _policySession,byte[] _operandB,short _offset,TPM_EO _operation)
     {
         authHandle = _authHandle;
         nvIndex = _nvIndex;
         policySession = _policySession;
         operandB = _operandB;
-        offset = (short)_offset;
+        offset = _offset;
         operation = _operation;
     }
     /**
