+++ TSS.Java/src/tss/tpm/TPMS_SIG_SCHEME_ECDAA.java
@@ -17,10 +17,10 @@ public class TPMS_SIG_SCHEME_ECDAA extends TpmStructure implements TPMU_SIG_SCHE
      * @param _hashAlg the hash algorithm used to digest the message 
      * @param _count the counter value that is used between TPM2_Commit() and the sign operation
      */
-    public TPMS_SIG_SCHEME_ECDAA(TPM_ALG_ID _hashAlg,int _count)
+    public TPMS_SIG_SCHEME_ECDAA(TPM_ALG_ID _hashAlg,short _count)
     {
         hashAlg = _hashAlg;
-        count = (short)_count;
+        count = _count;
     }
     /**
     * Most of the ECC signature schemes only require a hash algorithm to complete the definition and can be typed as TPMS_SCHEME_HASH. Anonymous algorithms also require a count value so they are typed to be TPMS_SCHEME_ECDAA.
